requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
arxiv==2.0.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.0
aiohttp==3.9.3